requirements.txt
streamlit
transformers
torch
pillow
pandas
altair